refactor: IPlayerState関連クラスをファイル分割
