+++ Assets/Scripts/Characters/Player/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25c4ae10f80524b3888a8a9dda910f0f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
