+++ Assets/Scripts/Characters/Player/State/AirState.cs
@@ -1,0 +1,5 @@
+public class AirState : PlayableState
+{
+    public AirState(Player player)
+        : base(player) { }
+}
