+++ Assets/Scripts/Characters/Player/State/DeadState.cs
@@ -1,0 +1,5 @@
+public class DeadState : UnplayableState
+{
+    public DeadState(Player player)
+        : base(player) { }
+}
