+++ Assets/Scripts/Characters/Player/State/DeadState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d3a726fd7d8b4049a3d16e184585eb95
