+++ Assets/Scripts/Characters/Player/State/FrozenState.cs
@@ -1,0 +1,5 @@
+public class FrozenState : UnplayableState
+{
+    public FrozenState(Player player)
+        : base(player) { }
+}
