+++ Assets/Scripts/Characters/Player/State/FrozenState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c88fa3b1b9b949f0a6d2be5af9d950d1
