+++ Assets/Scripts/Characters/Player/State/GoalState.cs
@@ -1,0 +1,5 @@
+public class GoalState : UnplayableState
+{
+    public GoalState(Player player)
+        : base(player) { }
+}
