+++ Assets/Scripts/Characters/Player/State/GoalState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ff855f724d0f45739eb9cebe27f80a72
