+++ Assets/Scripts/Characters/Player/State/GroundState.cs
@@ -1,0 +1,5 @@
+public class GroundState : PlayableState
+{
+    public GroundState(Player player)
+        : base(player) { }
+}
