+++ Assets/Scripts/Characters/Player/State/PlayableState.cs
@@ -1,0 +1,5 @@
+public abstract class PlayableState : PlayerStateBase
+{
+    protected PlayableState(Player player)
+        : base(player) { }
+}
