+++ Assets/Scripts/Characters/Player/State/PlayableState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0d83f87a6b5f434cb0f22923d4f3fd46
