+++ Assets/Scripts/Characters/Player/State/PlayerStateBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 872ac0f271fd4ff8b0cc2abf8d3121a1
