+++ Assets/Scripts/Characters/Player/State/UnplayableState.cs
@@ -1,0 +1,5 @@
+public abstract class UnplayableState : PlayerStateBase
+{
+    protected UnplayableState(Player player)
+        : base(player) { }
+}
