+++ Assets/Scripts/Characters/Player/State/UnplayableState.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1eff1b6745464f618e77ee100ba1aaf8
